backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cssselect2==0.8.0
django==4.2.21
django-cors-headers==4.7.0
django-vite==3.1.0
django-vite-plugin==4.1.0
djangorestframework==3.16.0
documents==0.0.1
python-dotenv
fonttools==4.58.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
importlib-metadata==8.7.0
markdown==3.8
markdown-it-py==3.0.0
markdown-pdf==1.7
mdurl==0.1.2
pandoc==2.4
pdfkit==1.0.0
pillow==11.2.1
plumbum==1.9.0
ply==3.11
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydyf==0.11.0
pymupdf==1.25.3
pypandoc==1.15
pyparsing==3.2.3
pyphen==0.17.2
pywin32==310
requests==2.32.3
rsa==4.9.1
sqlalchemy==2.0.41
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zipp==3.22.0
zopfli==0.2.3.post1
jinja2
jupyter
numpy
ollama
pandas
pyarrow
pydantic-ai
